requirements.txt
alembic==1.8.0
asyncpg==0.26.0
fastapi==0.78.0
motor==3.0.0
psycopg2-binary==2.9.3
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.2.0
SQLAlchemy==1.4.39
websockets==10.3
Werkzeug==2.1.2
uvicorn